+++ internal/bus/message_pipe_test.go
@@ -1,3 +1,9 @@
+/**
+ * Copyright (c) F5, Inc.
+ *
+ * This source code is licensed under the Apache License, Version 2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
 package bus
 
 import (
